Add new responsive renderer component
The responsive renderer component takes 4 props, renderDefault, renderDesktop, renderTablet, renderMobile, and renders based on the width of the client window. While JS is loading the component will render whats inside renderDefault, and this prop is optional if you don't want to render nothing until you have the size of the window. Noscript should be added on the renderDefault function if needed.

Also adds high order component to listen for resize events.