Replace withWindowDimensions HOC with a useWindowDimensions hook
Check hooks API for more info! I tried to replace this HOC with a Hook as it was a popular one that people were re-implementing in their projects.

This is merely exploratory, no need to rush and approve this but at least you can see it in action.

As for hooks vs classes debate, I have no opinion yet. Only thing I can say is that long term hooks can be less "bug-prone" as they don't deal with lifecycle events and are a more functional approach to side effects. They encapsulate effects nicely and I think it's something we should strive to start using more, as the React team is heading towards a more function-oriented approach instead of using classes.