Setup Forestry CMS to create authors and blog posts
Sets up Forestry CMS to manage markdown and authors.

I had to change some things in order for the blog posts to work. I had to remove our old folder structure. Now we should have only markdown files on `posts`, images on `images` and authors on `authors`.

I also set up the `JSON` transformer on Gatsby, so we can query authors directly.

Due to limitations on Gatsby, I had to customize `gatsby-node.js` to convert our paths on the frontmatter (author and featuredImage) from absolute to relative directly. To convert regular images on the markdown we use a plugin `gatsby-remark-relative-images` to handle that for us.

Also adds the new Reading List Blogpost!